Blurry placeholder for photo like in Medium added
Optimization for stateless components implmented. This brings
with it a significant improve in performance that the simply
use of stateless components couldn't achieve, it just as ease
as calling these components as they are, just functions, with
this we get to avoid calling render when is not need it.

Was a bit tricky but also a pleasure to mimmick the blurry
placeholder images that use Medium, i'm sure that for another
project needs  to get refactored for  sakes of modularity, but
right now it's doing its jobs very well.